requirements.txt
Flask==2.3.2
google-genai==1.50.0
faiss-cpu==1.7.4
numpy==1.26.0
pdfplumber==0.9.0
python-dotenv==1.0.0
